LayoutTests/svg/dom/SVGViewSpec-invalid-ref-crash-expected.txt

Confirm that null is returned when an SVGViewSpec is used after its corresponding element has been removed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL viewPreserveAspectRatio = svgView.preserveAspectRatio; should be null. Was [object SVGAnimatedPreserveAspectRatio].
FAIL viewTransform = svgView.transform; should be null. Was [object SVGTransformList].
FAIL viewViewBox = svgView.viewBox; should be null. Was [object SVGAnimatedRect].
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
